retire bespoke SCF drivers; rename OOO drivers to canonical names
The OpenOrbitalOptimizer-based drivers now cover every static field /
confinement / symmetry-averaging / frozen-occupation feature the
bespoke drivers had (PRs #177-181). Retire the bespoke SCF drivers
and promote the OOO variants to the canonical names:

- src/atomic/ooo.cpp   -> src/atomic/main.cpp        (binary: atomic)
- src/diatomic/ooo.cpp -> src/diatomic/main.cpp      (binary: diatomic)
- src/sadatom/ooo.cpp  -> src/sadatom/main.cpp       (binary: gensap)

Downstream scripts, examples/, docs, and the install(TARGETS ...)
list keep working unchanged -- `atomic`, `diatomic`, `gensap` refer
to the OOO-based drivers now.

CMakeLists updates: drop the old `add_executable(atomic_ooo ...)`,
`sadatom_ooo`, `diatomic_ooo` targets and link the OOO dependency
into the canonical binaries directly.

Net: -3021 lines across the three main.cpp files.

Not retired here:
- src/sadatom/solver.{cpp,h} and src/sadatom/configurations.{cpp,h}
  provide the bespoke sadatom SCFSolver as an atomic-orbital-guess
  helper for src/diatomic/twodquadrature.cpp (used by the
  diatomic_cbasis / diatomic_cpl analysis tools). Kept intact.
- src/general/diis.{cpp,h} and src/general/lbfgs.{cpp,h} kept for
  the same reason -- solver.cpp uses them.
