+++ src/diatomic/main.cpp
@@ -12,1101 +12,504 @@
  * See the LICENSE file at the root of this source distribution
  * for the full license text.
  */
+
+// Diatomic driver using OpenOrbitalOptimizer. Supports:
+//   - Restricted (closed-shell) DFT: one particle type, per-symmetry blocks,
+//     max_occupation = 2 per orbital.
+//   - Unrestricted DFT: two particle types (alpha, beta) each split into the
+//     same symmetry blocks; max_occupation = 1 per orbital.
+//
+// Symmetry decomposition uses basis.get_sym_idx(--symmetry) with 0/1/2 for
+// none / per-m / per-(m, parity). Per-block Sinvh_k is the symmetric
+// orthonormalization of S restricted to that block. Fock builder assembles
+// the AO Fock matrix once, then extracts and orthonormalizes per block.
+//
+// nela/nelb are derived from --Q and --M (spin multiplicity, 2S+1) via
+// scf::parse_nela_nelb, matching the bespoke diatomic driver's CLI.
+//
+// HF (x_func == 0 && c_func == 0) is deferred; picking any HF method throws.
+
 #include "../general/cmdline.h"
-#include "../general/checkpoint.h"
 #include "../general/constants.h"
-#include "../general/diis.h"
 #include "../general/dftfuncs.h"
-#include "../general/timer.h"
-#include "utils.h"
 #include "../general/elements.h"
 #include "../general/scf_helpers.h"
-#include "../general/scf_driver_common.h"
-#include "../general/model_potential.h"
+#include "../general/timer.h"
+
+#include "openorbitaloptimizer/scfsolver.hpp"
+
+#include "utils.h"
 #include "basis.h"
 #include "dftgrid.h"
 #include "twodquadrature.h"
+#include "../atomic/basis.h"
+#include <ArmaEigen.h>
+#include <Eigen/Eigenvalues>
 #include <cfloat>
 #include <climits>
+#include <sstream>
 
 using namespace helfem;
-
-void classify_orbitals(const arma::mat & C, const arma::ivec & mvals, const std::vector<arma::uvec> & mposidx, const std::vector<arma::uvec> & mnegidx) {
-  for(size_t io=0;io<C.n_cols;io++) {
-    arma::vec orb(C.col(io));
-
-    arma::vec opchar(mvals.n_elem), onchar(mvals.n_elem);
-    for(size_t c=0;c<mvals.n_elem;c++) {
-      opchar(c)=arma::norm(orb(mposidx[c]),"fro");
-      onchar(c)=arma::norm(orb(mnegidx[c]),"fro");
-    }
-
-    // Total character is
-    arma::vec ochar(opchar+onchar);
-
-    // Normalize
-    opchar/=arma::sum(ochar);
-    onchar/=arma::sum(ochar);
-    ochar/=arma::sum(ochar);
-
-    // Orbital symmetry is then
-    arma::uword oidx;
-    double stot=ochar.max(oidx);
-
-    // Symmetry threshold
-    double thr=0.999;
-
-    // Classification
-    std::ostringstream cl;
-
-    if(stot>=thr) {
-      // Orbital symmetry
-      char sym=' ';
-      if(opchar(oidx)>=thr)
-        sym='g';
-      else if(onchar(oidx)>=thr)
-        sym='u';
-
-      printf("Orbital %2i: m=%+i %c\n",(int) io+1,(int) mvals(oidx),sym);
-    } else {
-      printf("Orbital %2i: unknown\n",(int) io+1);
-    }
-  }
-}
-
-// normalize_matrix, gram_schmidt, report_ortho_deviation and
-// report_halfoverlap_error moved to helfem::scf_driver in
-// ../general/scf_driver_common.h.
-using helfem::scf_driver::normalize_matrix;
 
 int main(int argc, char **argv) {
   cmdline::parser parser;
 
-  // full option name, no short option, description, argument required
   parser.add<std::string>("Z1", 0, "first nuclear charge", true);
   parser.add<std::string>("Z2", 0, "second nuclear charge", true);
   parser.add<double>("Rbond", 0, "internuclear distance", true);
-  parser.add<bool>("angstrom", 0, "input distances in angstrom", false, false);
-  parser.add<int>("nela", 0, "number of alpha electrons", false, 0);
-  parser.add<int>("nelb", 0, "number of beta  electrons", false, 0);
   parser.add<int>("Q", 0, "charge state", false, 0);
-  parser.add<int>("M", 0, "spin multiplicity", false, 0);
-  parser.add<std::string>("lmax", 0, "maximum l quantum number", true, "");
+  parser.add<int>("M", 0, "spin multiplicity (2S+1); mutually exclusive with nela/nelb", false, 0);
+  parser.add<int>("nela", 0, "number of alpha electrons (leave 0 to derive from Q/M)", false, 0);
+  parser.add<int>("nelb", 0, "number of beta electrons (leave 0 to derive from Q/M)", false, 0);
+  parser.add<std::string>("lmax", 0, "maximum l quantum number (single int if mmax given, else comma list)", true, "");
   parser.add<int>("mmax", 0, "maximum m quantum number", false, -1);
   parser.add<double>("Rmax", 0, "practical infinity in au", false, 40.0);
-  parser.add<int>("grid", 0, "type of grid: 1 for linear, 2 for quadratic, 3 for polynomial, 4 for exponential", false, 4);
+  parser.add<int>("grid", 0, "type of grid: 1 linear, 2 quadratic, 3 polynomial, 4 exponential", false, 4);
   parser.add<double>("zexp", 0, "parameter in radial grid", false, 1.0);
   parser.add<int>("nelem", 0, "number of elements", true);
   parser.add<int>("nnodes", 0, "number of nodes per element", false, 15);
   parser.add<int>("nquad", 0, "number of quadrature points", false, 0);
-  parser.add<int>("maxit", 0, "maximum number of iterations", false, 50);
-  parser.add<double>("convthr", 0, "convergence threshold", false, 1e-7);
-  parser.add<double>("Ez", 0, "electric dipole field", false, 0.0);
-  parser.add<double>("Qzz", 0, "electric quadrupole field", false, 0.0);
-  parser.add<double>("Bz", 0, "magnetic dipole field", false, 0.0);
-  parser.add<bool>("diag", 0, "exact diagonalization", false, 1);
-  parser.add<int>("finitenuc", 0, "finite nuclear model", false, false);
-  parser.add<double>("Rrms1", 0, "nucleus 1 radius", false, 0.0);
-  parser.add<double>("Rrms2", 0, "nucleus 2 radius", false, 0.0);
-  parser.add<std::string>("method", 0, "method to use", false, "HF");
+  parser.add<std::string>("method", 0, "DFT method to use", false, "lda_x");
   parser.add<int>("ldft", 0, "theta rule for dft quadrature (0 for auto)", false, 0);
   parser.add<int>("mdft", 0, "phi rule for dft quadrature (0 for auto)", false, 0);
   parser.add<double>("dftthr", 0, "density threshold for dft", false, 1e-12);
-  parser.add<int>("restricted", 0, "spin-restricted orbitals", false, -1);
-  parser.add<int>("symmetry", 0, "force orbital symmetry", false, 1);
   parser.add<int>("primbas", 0, "primitive radial basis", false, 4);
-  parser.add<double>("diiseps", 0, "when to start mixing in diis", false, 1e-2);
-  parser.add<double>("diisthr", 0, "when to switch over fully to diis", false, 1e-3);
-  parser.add<int>("diisorder", 0, "length of diis history", false, 5);
-  parser.add<int>("readocc", 0, "read occupations from file, use until nth build", false, 0);
-  parser.add<double>("perturb", 0, "randomly perturb initial guess", false, 0.0);
-  parser.add<int>("seed", 0, "seed for random perturbation", false, 0);
-  parser.add<int>("iguess", 0, "guess: 0 for core, 1 for GSZ, 2 for SAP, 3 for TF", false, 2);
-  parser.add<std::string>("load", 0, "load guess from checkpoint", false, "");
-  parser.add<std::string>("save", 0, "save calculation to checkpoint", false, "helfem.chk");
+  parser.add<int>("restricted", 0, "spin-restricted: 1 restricted, 0 unrestricted, -1 auto from nela/nelb", false, -1);
+  parser.add<int>("symmetry", 0, "orbital symmetry: 0 none, 1 per-m, 2 per-(m,parity) (homonuclear only)", false, 1);
   parser.add<std::string>("x_pars", 0, "file for parameters for exchange functional", false, "");
   parser.add<std::string>("c_pars", 0, "file for parameters for correlation functional", false, "");
+
+  // External static fields (parity with bespoke diatomic driver).
+  parser.add<double>("Ez",  0, "electric dipole field",     false, 0.0);
+  parser.add<double>("Qzz", 0, "electric quadrupole field", false, 0.0);
+  parser.add<double>("Bz",  0, "magnetic dipole field",     false, 0.0);
+
+  // Fock symmetry averaging over m values (parity with bespoke diatomic).
   parser.add<bool>("maverage", 0, "average Fock matrix over m values", false, false);
+
+  // Frozen per-block occupations from occs.dat (parity with bespoke
+  // diatomic's --readocc, but a bool since OOO's fixed-per-block
+  // occupations apply for the whole SCF, not just N Fock builds).
+  parser.add<bool>("readocc", 0, "read frozen per-block occupations from occs.dat", false, false);
+
   parser.parse_check(argc, argv);
 
-  // Get parameters
-  double Rmax(parser.get<double>("Rmax"));
-  int igrid(parser.get<int>("grid"));
-  double zexp(parser.get<double>("zexp"));
-  double Ez(parser.get<double>("Ez"));
-  double Qzz(parser.get<double>("Qzz"));
-  double Bz(parser.get<double>("Bz"));
+  const int Z1        = get_Z(parser.get<std::string>("Z1"));
+  const int Z2        = get_Z(parser.get<std::string>("Z2"));
+  const double Rbond  = parser.get<double>("Rbond");
+        int Q         = parser.get<int>("Q");
+        int M         = parser.get<int>("M");
+        int nela      = parser.get<int>("nela");
+        int nelb      = parser.get<int>("nelb");
+  const std::string lmax_str = parser.get<std::string>("lmax");
+        int mmax      = parser.get<int>("mmax");
+  const double Rmax   = parser.get<double>("Rmax");
+  const int igrid     = parser.get<int>("grid");
+  const double zexp   = parser.get<double>("zexp");
+  const int Nelem     = parser.get<int>("nelem");
+  const int Nnodes    = parser.get<int>("nnodes");
+        int Nquad     = parser.get<int>("nquad");
+  const std::string method = parser.get<std::string>("method");
+  const int ldft_arg  = parser.get<int>("ldft");
+  const int mdft_arg  = parser.get<int>("mdft");
+  const double dftthr = parser.get<double>("dftthr");
+  const int primbas   = parser.get<int>("primbas");
+        int restr     = parser.get<int>("restricted");
+        int symm      = parser.get<int>("symmetry");
+  const std::string xparf = parser.get<std::string>("x_pars");
+  const std::string cparf = parser.get<std::string>("c_pars");
+  const double Ez     = parser.get<double>("Ez");
+  const double Qzz    = parser.get<double>("Qzz");
+  const double Bz     = parser.get<double>("Bz");
+  const bool   maverage = parser.get<bool>("maverage");
+  const bool   readocc  = parser.get<bool>("readocc");
 
-  int maxit(parser.get<int>("maxit"));
-  double convthr(parser.get<double>("convthr"));
+  // Derive nela/nelb from Q, M -- same convention as the bespoke diatomic
+  // driver. Total nuclear charge is Z1 + Z2.
+  scf::parse_nela_nelb(nela, nelb, Q, M, Z1 + Z2);
+  if (restr == -1) restr = (nela == nelb) ? 1 : 0;
+  const bool restricted = (restr != 0);
+  if (restricted && nela != nelb)
+    throw std::logic_error("Restricted mode requires nela == nelb (closed shell). "
+                            "Use --restricted=0 (or leave -1 for auto) for open-shell.");
+  const int Ntot = nela + nelb;
 
-  bool diag(parser.get<bool>("diag"));
-  int restr(parser.get<int>("restricted"));
-  int symm(parser.get<int>("symmetry"));
-  int iguess(parser.get<int>("iguess"));
-  bool maverage(parser.get<bool>("maverage"));
+  arma::vec x_pars, c_pars;
+  if (xparf.size()) x_pars = helfem::to_arma(scf::parse_xc_params(xparf));
+  if (cparf.size()) c_pars = helfem::to_arma(scf::parse_xc_params(cparf));
 
-  int primbas(parser.get<int>("primbas"));
-  // Number of elements
-  int Nelem(parser.get<int>("nelem"));
-  // Number of nodes
-  int Nnodes(parser.get<int>("nnodes"));
-  // Order of quadrature rule
-  int Nquad(parser.get<int>("nquad"));
-  // Angular grid
-  std::string lmax(parser.get<std::string>("lmax"));
-  int mmax(parser.get<int>("mmax"));
-
-  // DFT angular grid
-  int ldft(parser.get<int>("ldft"));
-  int mdft(parser.get<int>("mdft"));
-  double dftthr(parser.get<double>("dftthr"));
-
-  // Nuclear charge
-  int Z1(get_Z(parser.get<std::string>("Z1")));
-  int Z2(get_Z(parser.get<std::string>("Z2")));
-  double Rbond(parser.get<double>("Rbond"));
-  // Number of occupied states
-  int nela(parser.get<int>("nela"));
-  int nelb(parser.get<int>("nelb"));
-  int Q(parser.get<int>("Q"));
-  int M(parser.get<int>("M"));
-
-  // Finite nucleus
-  int finitenuc(parser.get<int>("finitenuc"));
-  double Rrms1(parser.get<double>("Rrms1"));
-  double Rrms2(parser.get<double>("Rrms2"));
-
-  double diiseps=parser.get<double>("diiseps");
-  double diisthr=parser.get<double>("diisthr");
-  int diisorder=parser.get<int>("diisorder");
-
-  std::string method(parser.get<std::string>("method"));
-
-  double perturb=parser.get<double>("perturb");
-  int seed=parser.get<int>("seed");
-
-  std::string save(parser.get<std::string>("save"));
-  std::string load(parser.get<std::string>("load"));
-
-  std::string xparf(parser.get<std::string>("x_pars"));
-  std::string cparf(parser.get<std::string>("c_pars"));
-
-  // Set parameters if necessary
-  arma::vec xpars, cpars;
-  if(xparf.size()) {
-    xpars = helfem::to_arma(scf::parse_xc_params(xparf));
-    xpars.t().print("Exchange functional parameters");
-  }
-  if(cparf.size()) {
-    cpars = helfem::to_arma(scf::parse_xc_params(cparf));
-    cpars.t().print("Correlation functional parameters");
-  }
-
-  // Open checkpoint in save mode
-  Checkpoint chkpt(save,true);
-
-  // Read occupations from file?
-  int readocc=parser.get<int>("readocc");
-  if(readocc<0)
-    readocc=INT_MAX;
-  arma::imat occs;
-  if(readocc) {
-    occs.load("occs.dat",arma::raw_ascii);
-    if(occs.n_cols < 3) {
-      throw std::logic_error("Must have at least three columns in occupation data.\n");
-    }
-  }
-
-  if(parser.get<bool>("angstrom")) {
-    // Convert to atomic units
-    Rbond*=ANGSTROMINBOHR;
-  }
-
-  scf::parse_nela_nelb(nela,nelb,Q,M,Z1+Z2);
-  if(restr==-1) {
-    // If number of electrons differs then unrestrict
-    restr=(nela==nelb);
-  }
-  chkpt.write("nela",nela);
-  chkpt.write("nelb",nelb);
-
-  std::vector<std::string> rcalc(2);
-  rcalc[0]="unrestricted";
-  rcalc[1]="restricted";
-
-  printf("Running %s %s calculation with Rmax=%e and %i elements.\n",rcalc[restr].c_str(),method.c_str(),Rmax,Nelem);
-
-  // Get primitive basis
-  auto poly(std::shared_ptr<const polynomial_basis::PolynomialBasis>(polynomial_basis::get_basis(primbas,Nnodes)));
-
-  if(Nquad==0)
-    // Set default value
-    Nquad=5*poly->get_nbf();
-  else if(Nquad<2*poly->get_nbf())
-    throw std::logic_error("Insufficient radial quadrature.\n");
-
-  printf("Using %i point quadrature rule.\n",Nquad);
-
-  arma::ivec lmmax;
-  if(mmax>=0) {
-    lmmax.ones(mmax+1);
-    lmmax*=atoi(lmax.c_str());
-  } else {
-    // Parse list of l values
-    std::vector<arma::uword> lmmaxv;
-    std::stringstream ss(lmax);
-    while( ss.good() ) {
-      std::string substr;
-      getline( ss, substr, ',' );
-      lmmaxv.push_back(atoi(substr.c_str()));
-    }
-    lmmax=arma::conv_to<arma::ivec>::from(lmmaxv);
-  }
-  // l and m values
-  arma::ivec lval, mval;
-  diatomic::basis::lm_to_l_m(lmmax,lval,mval);
-
-  double Rhalf(0.5*Rbond);
-  double mumax(utils::arcosh(Rmax/Rhalf));
-  arma::vec bval(atomic::basis::normal_grid(Nelem, mumax, igrid, zexp));
-
-  diatomic::basis::TwoDBasis basis(Z1, Z2, Rhalf, poly, Nquad, bval, lval, mval);
-  chkpt.write(basis);
-  printf("Basis set consists of %i angular shells composed of %i radial functions, totaling %i basis functions\n",(int) basis.Nang(), (int) basis.Nrad(), (int) basis.Nbf());
-
-  printf("One-electron matrix requires %s\n",scf::memory_size(basis.mem_1el()).c_str());
-  printf("Auxiliary one-electron integrals require %s\n",scf::memory_size(basis.mem_1el_aux()).c_str());
-  printf("Auxiliary two-electron integrals require %s\n",scf::memory_size(basis.mem_2el_aux()).c_str());
-
-  double Enucr=Z1*Z2/Rbond;
-  chkpt.write("Enucr",Enucr);
-  printf("Left- and right-hand nuclear charges are %i and %i at distance % .3f\n",Z1,Z2,Rbond);
-  printf("Nuclear repulsion energy is %e\n",Enucr);
-  printf("Number of electrons is %i %i\n",nela,nelb);
-
-  // Collect basis function indices
-  arma::ivec mvals;
-  {
-    arma::ivec mv(basis.get_m());
-    arma::uvec idx(arma::find_unique(mv,true));
-    mvals=mv(idx);
-  }
-  std::vector<arma::uvec> midx(mvals.n_elem), mposidx(mvals.n_elem), mnegidx(mvals.n_elem);
-  for(size_t i=0;i<midx.size();i++) {
-    midx[i]=basis.m_indices(mvals(i));
-    mposidx[i]=basis.m_indices(mvals(i),false);
-    mnegidx[i]=basis.m_indices(mvals(i),true);
-  }
-
-  // Symmetry indices
-  std::vector<arma::uvec> dsym;
-  if(symm==2 && Z1!=Z2) {
-    printf("Warning - asked for homonuclear symmetry for heteronuclear molecule. Relaxing restriction.\n");
-    symm=1;
-  }
-  if(symm==2 && (Ez!=0.0 || Qzz!=0.0)) {
-    printf("Warning - asked for full orbital symmetry in presence of electric field. Relaxing restriction.\n");
-    symm=1;
-  }
-  if(symm==2 && Bz!=0.0) {
-    printf("Warning - asked for full orbital symmetry in presence of magnetic field. Relaxing restriction.\n");
-    symm=1;
-  }
-  if(symm)
-    dsym=basis.get_sym_idx(symm);
-
-  // For m averaging
-  std::vector< std::vector<arma::uvec> > mavg_idx;
-  for(int m=0;m<=arma::max(arma::abs(basis.get_mval()));m++) {
-    // Construct set of indices to average over
-    std::vector<arma::uvec> entry;
-    entry.push_back(basis.m_indices(m));
-    if(m>0)
-      entry.push_back(basis.m_indices(-m));
-    mavg_idx.push_back(entry);
-  }
-
-  // Forced occupations?
-  arma::ivec occnuma, occnumb;
-  std::vector<arma::uvec> occsym;
-  if(readocc) {
-    // Number of occupied alpha orbitals is first column
-    occnuma=occs.col(0);
-    // Number of occupied beta orbitals is second column
-    occnumb=occs.col(1);
-
-    // Heteronuclear molecule?
-    if(Z1 != Z2 && occs.n_cols != 3)
-      throw std::logic_error("Heteronuclear molecule orbital occupations must have three columns.\n");
-    if(Z1 == Z2 && (occs.n_cols != 3) && (occs.n_cols != 4))
-      throw std::logic_error("Homonuclear molecule orbital occupations must have three or four columns.\n");
-
-    if(occs.n_cols == 3) {
-      // m value is third column
-      for(size_t i=0;i<occs.n_rows;i++)
-        occsym.push_back(basis.m_indices(occs(i,2)));
-    } else if(occs.n_cols == 4) {
-      if(symm!=2)
-        throw std::logic_error("For use of homonuclear orbital occupations, must turn on use of full symmetry.\n");
-      // m value is third column, parity is fourth column
-      for(size_t i=0;i<occs.n_rows;i++) {
-        if(occs(i,3)!=1 && occs(i,3)!=-1) {
-          std::ostringstream oss;
-          oss << "Error on line " << i+1 << " of orbital occupations: parity must be +1 or -1\n";
-          throw std::logic_error(oss.str());
-        }
-        occsym.push_back(basis.m_indices(occs(i,2),(occs(i,3)==-1)));
-      }
-    }
-
-    // Check consistency of values
-    if(arma::sum(occnuma) != nela) {
-      std::ostringstream oss;
-      oss << "Specified alpha occupations don't match wanted spin state.\n";
-      oss << "Occupying " << arma::sum(occnuma) << " orbitals but should have " << nela << " orbitals.\n";
-      throw std::logic_error(oss.str());
-    }
-    if(arma::sum(occnumb) != nelb) {
-      std::ostringstream oss;
-      oss << "Specified alpha occupations don't match wanted spin state.\n";
-      oss << "Occupying " << arma::sum(occnumb) << " orbitals but should have " << nelb << " orbitals.\n";
-      throw std::logic_error(oss.str());
-    }
-  }
-
-  // Functional
   int x_func, c_func;
   ::parse_xc_func(x_func, c_func, method);
   ::print_info(x_func, c_func);
-  if(!is_supported(x_func))
-    throw std::logic_error("The specified exchange functional is not currently supported in HelFEM.\n");
-  if(!is_supported(c_func))
-    throw std::logic_error("The specified correlation functional is not currently supported in HelFEM.\n");
+  if (!is_supported(x_func) || !is_supported(c_func))
+    throw std::logic_error("The specified functional is not supported in HelFEM.\n");
 
-  bool dft=(x_func>0 || c_func>0);
-  if(is_range_separated(x_func))
-    throw std::logic_error("Range separated functionals are not supported.\n");
-  // Fraction of exact exchange
-  double kfrac(exact_exchange(x_func));
-  if(kfrac!=0.0)
-    printf("\nUsing hybrid exchange with % .3f %% of exact exchange.\n",kfrac*100);
-  else
-    printf("\nA pure exchange functional used, no exact exchange.\n");
+  double kfrac, kshort, omega;
+  range_separation(x_func, omega, kfrac, kshort);
+  const bool have_exx = (kfrac != 0.0 || kshort != 0.0);
+  const bool have_xc  = (x_func != 0 || c_func != 0);
+  if (have_exx && kshort != 0.0)
+    throw std::logic_error("Range-separated hybrids not yet supported in diatomic_ooo (needs omega wiring).\n");
 
-  Timer timer;
+  auto poly = std::shared_ptr<const polynomial_basis::PolynomialBasis>(
+      polynomial_basis::get_basis(primbas, Nnodes));
+  if (Nquad == 0) Nquad = 5 * poly->get_nbf();
+  else if (Nquad < 2 * poly->get_nbf())
+    throw std::logic_error("Insufficient radial quadrature.\n");
 
-  // Form overlap matrix
-  arma::mat S(basis.overlap());
-  chkpt.write("S",S);
-  // Form kinetic energy matrix
-  arma::mat T(basis.kinetic());
-  chkpt.write("T",T);
-
-  helfem::diatomic::dftgrid::DFTGrid grid;
-  if(dft) {
-    if(ldft==0)
-      // Default value: we have 2*lmax from the bra and ket and 2 from
-      // the volume element, and allow for 2*lmax from the
-      // density/potential. Add in 10 more for a bit more accuracy.
-      ldft=4*arma::max(lmmax)+12;
-    if(ldft<(int) (2*arma::max(lmmax)+2))
-      throw std::logic_error("Increase ldft to guarantee accuracy of quadrature!\n");
-
-    if(mdft==0)
-      // Default value: we have 2*mmax from the bra and ket, and allow
-      // for 2*mmax from the density/potential. Add in 5 to make
-      // sure quadrature is still accurate for mmax=0
-      mdft=4*lmmax.n_elem+5;
-    if(mdft<(int) (2*lmmax.n_elem)) {
-      std::ostringstream oss;
-      oss << "Increase mdft at least to " << 2*lmmax.n_elem << " to guarantee accuracy of quadrature!\n";
-      throw std::logic_error(oss.str());
-    }
-
-    // Form grid
-    grid=helfem::diatomic::dftgrid::DFTGrid(&basis,ldft,mdft);
-
-    // Basis function norms
-    arma::vec bfnorm(arma::pow(arma::diagvec(S),-0.5));
-
-    // Check accuracy of grid
-    double Sthr=1e-10;
-    double Tthr=1e-8;
-    bool inacc=false;
-    {
-      arma::mat Sdft(grid.eval_overlap());
-      Sdft-=S;
-      normalize_matrix(Sdft,bfnorm);
-
-      double Serr(arma::norm(Sdft,"fro"));
-      printf("Error in overlap matrix evaluated through xc grid is %e\n",Serr);
-      fflush(stdout);
-      if(Serr>=Sthr)
-        inacc=true;
-    }
-    {
-      arma::mat Tdft(grid.eval_kinetic());
-      // Compute relative error
-      for(size_t j=0;j<Tdft.n_cols;j++)
-        for(size_t i=0;i<Tdft.n_rows;i++)
-          Tdft(i,j)=std::abs(Tdft(i,j)-T(i,j))/(1+std::abs(T(i,j)));
-
-      double Terr(arma::norm(Tdft,"fro"));
-      printf("Relative error in kinetic matrix evaluated through xc grid is %e\n",Terr);
-      fflush(stdout);
-      if(Terr>=Tthr)
-        inacc=true;
-    }
-    if(inacc)
-      printf("Warning - possibly inaccurate quadrature!\n");
+  arma::ivec lmmax;
+  if (mmax >= 0) {
+    lmmax.ones(mmax + 1);
+    lmmax *= std::atoi(lmax_str.c_str());
+  } else {
+    std::vector<arma::uword> lmmaxv;
+    std::stringstream ss(lmax_str);
+    std::string tok;
+    while (std::getline(ss, tok, ','))
+      lmmaxv.push_back(std::atoi(tok.c_str()));
+    lmmax = arma::conv_to<arma::ivec>::from(lmmaxv);
+    mmax = static_cast<int>(lmmaxv.size()) - 1;
   }
 
-  // Get half-inverse
-  timer.set();
-  arma::mat Sinvh(basis.Sinvh(!diag,symm));
-  chkpt.write("Sinvh",Sinvh);
-  printf("Half-inverse formed in %.6f\n",timer.get());
-  helfem::scf_driver::report_ortho_deviation(S, Sinvh);
-  arma::mat Sh(basis.Shalf(!diag,symm));
-  chkpt.write("Sh",Sh);
-  printf("Half-overlap formed in %.6f\n",timer.get());
-  helfem::scf_driver::report_halfoverlap_error(Sh, Sinvh);
+  arma::ivec lval, mval;
+  diatomic::basis::lm_to_l_m(lmmax, lval, mval);
 
-  // Form nuclear attraction energy matrix
-  Timer tnuc;
-  arma::mat Vnuc;
-  if(finitenuc==0)
-    Vnuc=basis.nuclear();
-  else {
-    modelpotential::ModelPotential *pot1(modelpotential::get_nuclear_model((modelpotential::nuclear_model_t) (finitenuc-1),Z1,Rrms1));
-    modelpotential::ModelPotential *pot2(modelpotential::get_nuclear_model((modelpotential::nuclear_model_t) (finitenuc-1),Z2,Rrms2));
-    int lquad = (ldft>0) ? ldft : 4*arma::max(lmmax)+12;
-    helfem::diatomic::twodquad::TwoDGrid qgrid;
-    qgrid=helfem::diatomic::twodquad::TwoDGrid(&basis,lquad);
+  const double Rhalf = 0.5 * Rbond;
+  const double mumax = utils::arcosh(Rmax / Rhalf);
+  arma::vec bval(atomic::basis::normal_grid(Nelem, mumax, igrid, zexp));
 
-    arma::mat Squad(qgrid.overlap());
-    Squad-=S;
-    arma::vec bfnorm(arma::pow(arma::diagvec(S),-0.5));
-    normalize_matrix(Squad,bfnorm);
-
-    double Serr(arma::norm(Squad,"fro"));
-    printf("Error in overlap matrix evaluated on two-dimensional grid is %e\n",Serr);
-    fflush(stdout);
-
-    Vnuc=qgrid.model_potential(pot1,pot2);
-    delete pot1;
-    delete pot2;
+  // Homonuclear g/u symmetry is not a valid decomposition for a
+  // heteronuclear molecule -- get_sym_idx(2) still splits by (m, l%2)
+  // but the SCF then sees an artificially restricted trial space and
+  // converges to a WRONG energy (silent in the bespoke driver too
+  // until relaxation to symm=1). Match the bespoke driver's warn +
+  // relax.
+  if (symm == 2 && Z1 != Z2) {
+    printf("Warning - asked for homonuclear symmetry for heteronuclear molecule. Relaxing to symmetry=1.\n");
+    symm = 1;
   }
-  chkpt.write("Vnuc",Vnuc);
-
-  // Dipole coupling
-  const arma::mat dip(basis.dipole_z());
-  chkpt.write("dip",dip);
-  // Quadrupole coupling
-  const arma::mat quad(basis.quadrupole_zz());
-  chkpt.write("quad",quad);
-
-  // Nuclear dipole and quadrupole
-  const double nucdip=(Z2-Z1)*Rhalf;
-  const double nucquad=(Z1+Z2)*Rhalf*Rhalf;
-
-  // Electric field coupling (minus sign cancels one from charge)
-  const arma::mat Vel(Ez*dip + Qzz*quad/3.0);
-  chkpt.write("Vel",Vel);
-  // Magnetic field coupling
-  arma::mat Vmag(basis.Bz_field(Bz));
-  chkpt.write("Vmag",Vmag);
-  const double Enucfield(-Ez*nucdip - Qzz*nucquad/3.0);
-
-  // Form Hamiltonian
-  const arma::mat H0(T+Vnuc+Vel+Vmag);
-  chkpt.write("H0",H0);
-
-  printf("One-electron matrices formed in %.6f\n",timer.get());
-
-  // Occupied and virtual orbitals
-  arma::mat Caocc, Cbocc, Cavirt, Cbvirt;
-  arma::vec Ea, Eb;
-  // Number of eigenenergies to print
-  arma::uword nena(std::min((arma::uword) nela+4,Sinvh.n_cols));
-  arma::uword nenb(std::min((arma::uword) nelb+4,Sinvh.n_cols));
-
-  // Guess orbitals
-  timer.set();
-  {
-    arma::mat Ca, Cb;
-    if(load.size()) {
-      printf("Guess orbitals from checkpoint\n");
-
-      // Load checkpoint
-      Checkpoint loadchk(load,false);
-      // Old basis set
-      diatomic::basis::TwoDBasis oldbasis;
-      loadchk.read(oldbasis);
-
-      arma::mat oldSinvh;
-      loadchk.read("Sinvh",oldSinvh);
-
-      // Interbasis overlap
-      arma::mat S12(basis.overlap(oldbasis));
-
-      switch(iguess) {
-      case(0):
-        printf("Guess orbitals from Fock matrix projection\n");
-	{
-	  // Convert to orthonormal basis
-	  S12=arma::trans(Sinvh)*S12*oldSinvh;
-	  // Helper
-	  arma::mat SSinvh(S*Sinvh);
-
-	  helfem::scf_driver::project_and_diagonalize(
-	      loadchk, "Fa", oldSinvh, S12, SSinvh, Sinvh, symm, dsym, Ea, Ca);
-	  helfem::scf_driver::project_and_diagonalize(
-	      loadchk, "Fb", oldSinvh, S12, SSinvh, Sinvh, symm, dsym, Eb, Cb);
-	}
-        break;
-
-      case(1):
-      default:
-        // Project lowest orbitals
-        printf("Guess orbitals from previous calculation\n");
-      {
-        // Projector
-        arma::mat P((Sinvh*arma::trans(Sinvh))*S12);
-
-        // Orbitals
-        arma::mat C;
-
-        // Alpha orbitals
-        loadchk.read("Ca",C);
-        // Project onto new basis: C1 = S11^-1 S12 C2
-        Ca=P*C;
-
-        // Beta orbitals
-        loadchk.read("Cb",C);
-        Cb=P*C;
-
-        // Run Gram-Schmidt to make sure orbitals are orthonormal
-        helfem::scf_driver::gram_schmidt(Ca, S, nela);
-        helfem::scf_driver::gram_schmidt(Cb, S, nelb);
-
-        // Read in orbital energies
-        loadchk.read("Ea",Ea);
-        if(Ea.n_elem<Ca.n_cols)
-          Ea=Ea.subvec(0,Ca.n_cols-1);
-        loadchk.read("Eb",Eb);
-        if(Eb.n_elem<Cb.n_cols)
-          Eb=Eb.subvec(0,Cb.n_cols-1);
-      }
-      break;
-      }
-    } else {
-      modelpotential::ModelPotential * p1, * p2;
-      switch(iguess) {
-      case(0):
-        // Use core guess
-        printf("Guess orbitals from core Hamiltonian\n");
-        p1 = new modelpotential::PointNucleus(Z1);
-        p2 = new modelpotential::PointNucleus(Z2);
-        break;
-
-      case(1):
-        // Use GSZ guess
-        printf("Guess orbitals from GSZ screened nucleus\n");
-        p1 = new modelpotential::GSZAtom(Z1);
-        p2 = new modelpotential::GSZAtom(Z2);
-        break;
-
-      case(2):
-        // Use SAP guess
-        printf("Guess orbitals from SAP screened nucleus\n");
-        p1 = new modelpotential::SAPAtom(Z1);
-        p2 = new modelpotential::SAPAtom(Z2);
-        break;
-
-      case(3):
-        // Use Thomas-Fermi guess
-        printf("Guess orbitals from Thomas-Fermi nucleus\n");
-        p1 = new modelpotential::TFAtom(Z1);
-        p2 = new modelpotential::TFAtom(Z2);
-        break;
-
-      default:
-        throw std::logic_error("Unsupported guess\n");
-      }
-
-      // Quadrature grid
-      int lquad = (ldft>0) ? ldft : 4*arma::max(lmmax)+12;
-      helfem::diatomic::twodquad::TwoDGrid qgrid;
-      qgrid=helfem::diatomic::twodquad::TwoDGrid(&basis,lquad);
-
-      arma::mat Squad(qgrid.overlap());
-      Squad-=S;
-      arma::vec bfnorm(arma::pow(arma::diagvec(S),-0.5));
-      normalize_matrix(Squad,bfnorm);
-
-      double Serr(arma::norm(Squad,"fro"));
-      printf("Error in overlap matrix evaluated on two-dimensional grid is %e\n",Serr);
-      fflush(stdout);
-
-      arma::mat Hguess(T+Vel+Vmag+qgrid.model_potential(p1,p2));
-      delete p1;
-      delete p2;
-
-      // Diagonalize
-      if(symm)
-        { // Phase 5.12 bridge: eig_gsym_sub is Eigen.
-          helfem::Vector Ea_e; helfem::Matrix Ca_e;
-          scf::eig_gsym_sub(Ea_e, Ca_e, helfem::to_eigen(Hguess), helfem::to_eigen(Sinvh), dsym);
-          Ea = helfem::to_arma(Ea_e);
-          Ca = helfem::to_arma(Ca_e);
-        }
-      else
-        { // Phase 5.11 bridge: eig_gsym is Eigen.
-          helfem::Vector Ea_e; helfem::Matrix Ca_e;
-          scf::eig_gsym(Ea_e, Ca_e, helfem::to_eigen(Hguess), helfem::to_eigen(Sinvh));
-          Ea = helfem::to_arma(Ea_e);
-          Ca = helfem::to_arma(Ca_e);
-        }
-
-      // Beta guess is the same as the alpha guess
-      Cb=Ca;
-      Eb=Ea;
-
-      // Enforce occupation according to specified symmetry
-      if(readocc) {
-	{ // Phase 5.13 bridge: enforce_occupations is Eigen.
-	  helfem::Matrix Ca_e = helfem::to_eigen(Ca);
-	  helfem::Vector Ea_e = helfem::to_eigen(Ea);
-	  scf::enforce_occupations(Ca_e, Ea_e, helfem::to_eigen(S), occnuma, occsym);
-	  Ca = helfem::to_arma(Ca_e);
-	  Ea = helfem::to_arma(Ea_e);
-	}
-	if(restr && nela==nelb)
-	  Cb=Ca;
-	else
-	  { // Phase 5.13 bridge: enforce_occupations is Eigen.
-	    helfem::Matrix Cb_e = helfem::to_eigen(Cb);
-	    helfem::Vector Eb_e = helfem::to_eigen(Eb);
-	    scf::enforce_occupations(Cb_e, Eb_e, helfem::to_eigen(S), occnumb, occsym);
-	    Cb = helfem::to_arma(Cb_e);
-	    Eb = helfem::to_arma(Eb_e);
-	  }
-      }
-    }
-
-    // Perturb guess
-    if(perturb) {
-      // Generate norb x norb rotation matrix
-      arma::arma_rng::set_seed(seed);
-      Ca*=helfem::to_arma(scf::perturbation_matrix(Ca.n_cols,perturb));
-      if(restr && nela==nelb) {
-        Cb=Ca;
-      } else {
-        Cb*=helfem::to_arma(scf::perturbation_matrix(Cb.n_cols,perturb));
-      }
-      printf("Guess orbitals perturbed by %e\n",perturb);
-    }
-
-    // Alpha orbitals
-    Caocc=Ca.cols(0,nela-1);
-    if(Ca.n_cols>(size_t) nela)
-      Cavirt=Ca.cols(nela,Ca.n_cols-1);
-
-    // Beta guess
-    if(nelb)
-      Cbocc=Cb.cols(0,nelb-1);
-    if(Cb.n_cols>(size_t) nelb)
-      Cbvirt=Cb.cols(nelb,Cb.n_cols-1);
-
-    Ea.subvec(0,nena-1).t().print("Alpha orbital energies");
-    Eb.subvec(0,nenb-1).t().print("Beta  orbital energies");
-
-    printf("\n");
-    printf("Alpha orbital symmetries\n");
-    classify_orbitals(Caocc,mvals,mposidx,mnegidx);
-    if(nelb>0) {
-      printf("\n");
-      printf("Beta orbital symmetries\n");
-      classify_orbitals(Cbocc,mvals,mposidx,mnegidx);
-    }
-    printf("\n");
+  // symm=2 (per-(m, parity)) also breaks in the presence of an external
+  // electric or magnetic field along z -- the field mixes parity or m
+  // sectors. Match the bespoke diatomic driver's relaxation to symm=1.
+  if (symm == 2 && (Ez != 0.0 || Qzz != 0.0)) {
+    printf("Warning - asked for full orbital symmetry in presence of electric field. Relaxing restriction.\n");
+    symm = 1;
   }
-  printf("Initial guess performed in %.6f\n",timer.get());
+  if (symm == 2 && Bz != 0.0) {
+    printf("Warning - asked for full orbital symmetry in presence of magnetic field. Relaxing restriction.\n");
+    symm = 1;
+  }
 
-  printf("Computing two-electron integrals\n");
-  fflush(stdout);
-  timer.set();
-  basis.compute_tei(kfrac!=0.0);
-  printf("Done in %.6f\n",timer.get());
+  diatomic::basis::TwoDBasis basis(Z1, Z2, Rhalf, poly, Nquad, bval, lval, mval);
+  const size_t Nbf = basis.Nbf();
+  printf("Basis set: %i angular shells x %i radial = %i basis functions\n",
+          (int) basis.Nang(), (int) basis.Nrad(), (int) Nbf);
+  printf("Mode: %s, symmetry=%d, nela=%d nelb=%d\n",
+          restricted ? "restricted" : "unrestricted", symm, nela, nelb);
 
-  double Ekin=0.0, Epot=0.0, Ecoul=0.0, Exx=0.0, Exc=0.0, Eefield=0.0, Emfield=0.0, Etot=0.0;
-  double Eold=0.0;
+  // Diatomic chemistry-layer methods are still arma-native.
+  const arma::mat S    = basis.overlap();
+  const arma::mat T    = basis.kinetic();
+  const arma::mat Vnuc = basis.nuclear();
 
-  bool usediis=true, useadiis=true, diiscomb=false;
-  uDIIS diis(S,Sinvh,diiscomb,usediis,diiseps,diisthr,useadiis,true,diisorder);
-  double diiserr;
+  // External static-field one-electron matrices. Diatomic has two
+  // nuclei at (0, 0, +/- Rhalf); nuclear dipole/quadrupole moments are
+  // non-zero and add a constant scalar Enucfield to the total energy.
+  //   nucdip  = (Z2 - Z1) * Rhalf       (points along +z from Z1->Z2)
+  //   nucquad = (Z1 + Z2) * Rhalf^2
+  //   Enucfield = -Ez * nucdip - Qzz * nucquad / 3
+  // Vel/Vmag are the electron-side matrix contributions folded into H0.
+  const arma::mat dip  = basis.dipole_z();
+  const arma::mat quad = basis.quadrupole_zz();
+  const arma::mat Vel  = Ez * dip + Qzz * quad / 3.0;
+  const arma::mat Vmag = basis.Bz_field(Bz);
+  const double nucdip  = (Z2 - Z1) * Rhalf;
+  const double nucquad = (Z1 + Z2) * Rhalf * Rhalf;
+  const double Enucfield = -Ez * nucdip - Qzz * nucquad / 3.0;
+  const bool have_efield = (Ez != 0.0 || Qzz != 0.0);
+  const bool have_bfield = (Bz != 0.0);
 
-  // Density matrices
-  arma::mat P, Pa, Pb;
-
-  for(int i=1;i<=maxit;i++) {
-    printf("\n**** Iteration %i ****\n\n",i);
-
-    // Form density matrix
-    Pa=helfem::to_arma(scf::form_density(helfem::to_eigen(Caocc),nela));
-    Pb=helfem::to_arma(scf::form_density(helfem::to_eigen(Cbocc),nelb));
-    if(Pb.n_rows == 0)
-      Pb.zeros(Pa.n_rows,Pa.n_cols);
-    P=Pa+Pb;
-
-    chkpt.write("P",P);
-    chkpt.write("Pa",Pa);
-    chkpt.write("Pb",Pb);
-
-    printf("Tr Pa = %f\n",arma::trace(Pa*S));
-    if(nelb)
-      printf("Tr Pb = %f\n",arma::trace(Pb*S));
-    fflush(stdout);
-
-    Ekin=arma::trace(P*T);
-    Epot=arma::trace(P*Vnuc);
-    Eefield=arma::trace(P*Vel);
-    Emfield=arma::trace(P*Vmag)-Bz/2.0*(nela-nelb);
-    chkpt.write("Ekin",Ekin);
-    chkpt.write("Epot",Epot);
-    chkpt.write("Eefield",Eefield);
-    chkpt.write("Emfield",Emfield);
-
-    // Form Coulomb matrix
-    timer.set();
-    arma::mat J(basis.coulomb(P));
-    double tJ(timer.get());
-    Ecoul=0.5*arma::trace(P*J);
-    printf("Coulomb energy %.10e % .6f\n",Ecoul,tJ);
-    fflush(stdout);
-    chkpt.write("J",J);
-    chkpt.write("Ecoul",Ecoul);
-
-    // Form exchange matrix
-    timer.set();
-    arma::mat Ka, Kb;
-    if(kfrac!=0.0) {
-      Ka=kfrac*basis.exchange(Pa);
-
-      if(nelb) {
-        if(restr && nela==nelb)
-          Kb=Ka;
-        else
-          Kb=kfrac*basis.exchange(Pb);
-      } else
-        Kb.zeros(Cbocc.n_rows,Cbocc.n_rows);
-      double tK(timer.get());
-      Exx=0.5*arma::trace(Pa*Ka);
-      if(Kb.n_rows == Pb.n_rows && Kb.n_cols == Pb.n_cols)
-        Exx+=0.5*arma::trace(Pb*Kb);
-      printf("Exchange energy %.10e % .6f\n",Exx,tK);
-    } else {
-      Exx=0.0;
+  // For diatomic m-averaging: each outer entry groups the (+m, -m) pair
+  // of BF-index sets so scf::fock_symmetry_average enforces
+  // f(+m) == f(-m). At m == 0 there is no partner so the group has a
+  // single member and the "average" is a no-op on that block. Matches
+  // src/diatomic/main.cpp lines 318..328.
+  std::vector<std::vector<arma::uvec>> mavg_idx;
+  if (maverage) {
+    const int mmax_bf = arma::max(arma::abs(basis.get_mval()));
+    for (int m = 0; m <= mmax_bf; ++m) {
+      std::vector<arma::uvec> entry;
+      entry.push_back(basis.m_indices(m));
+      if (m > 0) entry.push_back(basis.m_indices(-m));
+      mavg_idx.push_back(entry);
     }
-    fflush(stdout);
+  }
 
-    chkpt.write("Ka",Ka);
-    chkpt.write("Kb",Kb);
-    chkpt.write("Exx",Exx);
+  // Symmetry decomposition.
+  std::vector<arma::uvec> dsym;
+  if (symm == 0) {
+    arma::uvec all(Nbf);
+    for (size_t i = 0; i < Nbf; ++i) all(i) = i;
+    dsym.push_back(all);
+  } else {
+    dsym = basis.get_sym_idx(symm);
+  }
+  const size_t nsym = dsym.size();
 
-    // Exchange-correlation
-    Exc=0.0;
+  // Per-block Sinvh_k.
+  std::vector<arma::mat> Sinvh_arma(nsym);
+  for (size_t k = 0; k < nsym; ++k) {
+    if (!dsym[k].n_elem) continue;
+    const arma::mat Sk = S(dsym[k], dsym[k]);
+    Sinvh_arma[k] = helfem::to_arma(scf::form_Sinvh(helfem::to_eigen(Sk), /*chol*/false));
+  }
+
+  const int lang = ldft_arg > 0 ? ldft_arg : 4 * arma::max(lval) + 12;
+  const int mang = mdft_arg > 0 ? mdft_arg : 4 * mmax + 12;
+  auto grid = helfem::diatomic::dftgrid::DFTGrid(&basis, lang, mang);
+  basis.compute_tei(have_exx);
+
+  const double Enucr = Z1 * Z2 / Rbond;
+
+  using OOO_Real = double;
+  const size_t nparttype = restricted ? 1 : 2;
+  OpenOrbitalOptimizer::IndexVector number_of_blocks_per_particle_type(nparttype);
+  Eigen::Matrix<OOO_Real, Eigen::Dynamic, 1> maximum_occupation(nsym * nparttype);
+  Eigen::Matrix<OOO_Real, Eigen::Dynamic, 1> number_of_particles(nparttype);
+  std::vector<std::string> block_descriptions;
+  block_descriptions.reserve(nsym * nparttype);
+
+  for (size_t t = 0; t < nparttype; ++t) {
+    number_of_blocks_per_particle_type(t) = static_cast<int>(nsym);
+    number_of_particles(t) = static_cast<OOO_Real>(restricted ? Ntot : (t == 0 ? nela : nelb));
+    for (size_t k = 0; k < nsym; ++k) {
+      maximum_occupation(t * nsym + k) = restricted ? 2.0 : 1.0;
+      block_descriptions.push_back(
+          (nparttype == 1 ? "" : (t == 0 ? "a:" : "b:")) + std::string("sym") + std::to_string(k));
+    }
+  }
+
+  // Accumulate a per-block density into the full-Nbf density matrix
+  // P_full through the block's basis-function scatter index dsym[k].
+  auto accumulate_density = [&](arma::mat & P_full, size_t k,
+                                 const helfem::Matrix & orb_e,
+                                 const helfem::Vector & occ_e) {
+    if (!dsym[k].n_elem) return;
+    const arma::mat orb  = helfem::to_arma(orb_e);
+    const arma::vec occ  = helfem::to_arma(occ_e);
+    const arma::mat C_k  = Sinvh_arma[k] * orb;
+    const arma::mat P_k  = C_k * arma::diagmat(occ) * C_k.t();
+    P_full(dsym[k], dsym[k]) += P_k;
+  };
+
+  // Extract F_full(dsym[k], dsym[k]), transform to the block's
+  // orthonormal basis via Sinvh_k^T . F_k . Sinvh_k, stash into fock[b].
+  auto orthonormalize_block =
+      [&](OpenOrbitalOptimizer::FockMatrix<OOO_Real> & fock, size_t b,
+          const arma::mat & F_full, size_t k) {
+    if (!dsym[k].n_elem) {
+      fock[b] = helfem::Matrix::Zero(0, 0);
+      return;
+    }
+    const arma::mat Fk_sub = F_full(dsym[k], dsym[k]);
+    const arma::mat F_orth = Sinvh_arma[k].t() * Fk_sub * Sinvh_arma[k];
+    fock[b] = helfem::to_eigen(F_orth);
+  };
+
+  OpenOrbitalOptimizer::FockBuilder<OOO_Real, OOO_Real> fock_builder =
+      [&](const OpenOrbitalOptimizer::DensityMatrix<OOO_Real, OOO_Real> & dm) {
+    const auto & orbitals    = dm.first;
+    const auto & occupations = dm.second;
+
+    // Density assembly. Restricted mode has a single closed-shell channel
+    // with max_occ = 2, so P comes straight from the alpha channel; no
+    // Pa/Pb split, no *0.5 double-scatter.
+    OpenOrbitalOptimizer::FockMatrix<OOO_Real> fock(nsym * nparttype);
+    arma::mat P(Nbf, Nbf, arma::fill::zeros);
+    arma::mat Pa, Pb;
+    double Exc = 0.0;
+    double nelnum = 0.0, ekin_grid = 0.0;
     arma::mat XCa, XCb;
-    if(dft) {
-      timer.set();
-      double nelnum;
-      double ekin;
-      if(restr && nela==nelb) {
-        grid.eval_Fxc(x_func, xpars, c_func, cpars, P, XCa, Exc, nelnum, ekin, dftthr);
-        XCb=XCa;
+
+    if (restricted) {
+      for (size_t k = 0; k < nsym; ++k)
+        accumulate_density(P, k, orbitals[k], occupations[k]);
+      if (have_xc)
+        grid.eval_Fxc(x_func, x_pars, c_func, c_pars, P, XCa, Exc, nelnum, ekin_grid, dftthr);
+      if (have_exx) Pa = 0.5 * P;
+    } else {
+      Pa.zeros(Nbf, Nbf);
+      Pb.zeros(Nbf, Nbf);
+      for (size_t k = 0; k < nsym; ++k) {
+        accumulate_density(Pa, k, orbitals[k],        occupations[k]);
+        accumulate_density(Pb, k, orbitals[nsym + k], occupations[nsym + k]);
+      }
+      P = Pa + Pb;
+      if (have_xc)
+        grid.eval_Fxc(x_func, x_pars, c_func, c_pars, Pa, Pb, XCa, XCb,
+                       Exc, nelnum, ekin_grid, nelb > 0, dftthr);
+    }
+
+    const double Ekin = arma::trace(P * T);
+    const double Enuc = arma::trace(P * Vnuc);
+    // External-field trace contributions. Vel/Vmag are zero when their
+    // coupling is off; Enucfield is a scalar independent of the density.
+    const double Eefield = have_efield ? arma::trace(P * Vel) + Enucfield : 0.0;
+    const double Emfield = have_bfield
+        ? arma::trace(P * Vmag) - 0.5 * Bz * (nela - nelb) : 0.0;
+
+    const arma::mat J = basis.coulomb(P);
+    const double Ecoul = 0.5 * arma::trace(P * J);
+
+    // HF exchange. See sign-convention note in atomic/ooo.cpp: basis.exchange
+    // returns a matrix that gets ADDED to the Fock and the energy
+    // contribution is +0.5*trace(Pspin*Kspin) per channel.
+    arma::mat Ka, Kb;
+    double Exx = 0.0;
+    if (have_exx) {
+      Ka = kfrac * basis.exchange(Pa);
+      Exx = 0.5 * arma::trace(Pa * Ka);
+      if (!restricted) {
+        Kb = kfrac * basis.exchange(Pb);
+        Exx += 0.5 * arma::trace(Pb * Kb);
       } else {
-        grid.eval_Fxc(x_func, xpars, c_func, cpars, Pa, Pb, XCa, XCb, Exc, nelnum, ekin, nelb>0, dftthr);
+        Exx *= 2.0;
       }
-      double txc(timer.get());
-      printf("DFT energy %.10e % .6f\n",Exc,txc);
-      printf("Error in integrated number of electrons % e\n",nelnum-nela-nelb);
-      if(ekin!=0.0)
-        printf("Error in integral of kinetic energy density % e\n",ekin-Ekin);
     }
+
+    const double Etot = Ekin + Enuc + Eefield + Emfield
+                       + Ecoul + Exc + Exx + Enucr;
+    printf("kinetic %.10f nuclear %.10f Enucr %.10f Coulomb %.10f XC %.10f Exx %.10f",
+            Ekin, Enuc, Enucr, Ecoul, Exc, Exx);
+    if (have_efield) printf(" Eefield %.10f", Eefield);
+    if (have_bfield) printf(" Emfield %.10f", Emfield);
+    printf("  total %.10f  (nel err %.3e)\n",
+            Etot, nelnum - static_cast<double>(Ntot));
     fflush(stdout);
 
-    chkpt.write("XCa",XCa);
-    chkpt.write("XCb",XCb);
-    chkpt.write("Exc",Exc);
-
-    // Fock matrices
-    arma::mat Fa(H0+J);
-    arma::mat Fb(H0+J);
-    if(Ka.n_rows == Fa.n_rows) {
-      Fa+=Ka;
-    }
-    if(Kb.n_rows == Fb.n_rows) {
-      Fb+=Kb;
-    }
-    if(dft) {
-      Fa+=XCa;
-      if(nelb>0) {
-        Fb+=XCb;
-      }
-    }
-    if(Bz!=0.0) {
-      // Add in the B*Sz term
-      Fa-=Bz*S/2.0;
-      Fb+=Bz*S/2.0;
-    }
-
-    // m averaging?
-    if(maverage) {
-      Fa=helfem::to_arma(scf::fock_symmetry_average(helfem::to_eigen(Fa),mavg_idx));
-      Fb=helfem::to_arma(scf::fock_symmetry_average(helfem::to_eigen(Fb),mavg_idx));
-    }
-    // Enforce symmetry of Fock matrix
-    if(symm) {
-      Fa=helfem::to_arma(scf::enforce_fock_symmetry(helfem::to_eigen(Fa),dsym));
-      Fb=helfem::to_arma(scf::enforce_fock_symmetry(helfem::to_eigen(Fb),dsym));
-    }
-
-    // ROHF update to Fock matrix
-    if(restr && nela!=nelb)
-      { // Phase 5.16 bridge: ROHF_update is Eigen.
-        helfem::Matrix Fa_e = helfem::to_eigen(Fa);
-        helfem::Matrix Fb_e = helfem::to_eigen(Fb);
-        scf::ROHF_update(Fa_e, Fb_e, helfem::to_eigen(P),
-                         helfem::to_eigen(Sh), helfem::to_eigen(Sinvh),
-                         nela, nelb);
-        Fa = helfem::to_arma(Fa_e);
-        Fb = helfem::to_arma(Fb_e);
-      }
-
-    chkpt.write("Fa",Fa);
-    chkpt.write("Fb",Fb);
-
-    // Update energy
-    Etot=Ekin+Epot+Eefield+Emfield+Ecoul+Exx+Exc+Enucr+Enucfield;
-    double dE=Etot-Eold;
-
-    printf("Total energy is % .10f\n",Etot);
-    if(i>1)
-      printf("Energy changed by %e\n",dE);
-    Eold=Etot;
-    fflush(stdout);
-
-    // Update DIIS
-    timer.set();
-    diis.update(Fa,Fb,Pa,Pb,Etot,diiserr);
-    printf("DIIS error is %e, update done in %.6f\n",diiserr,timer.get());
-    fflush(stdout);
-
-    // Solve DIIS to get Fock update
-    timer.set();
-    diis.solve_F(Fa,Fb);
-    printf("DIIS solution done in %.6f\n",timer.get());
-    fflush(stdout);
-
-    // Have we converged? Note that DIIS error is still wrt full space, not active space.
-    bool convd=(diiserr<convthr) && (std::abs(dE)<convthr);
-
-    // Diagonalize Fock matrix to get new orbitals
-    timer.set();
-    arma::mat Ca, Cb;
-    if(symm)
-      { // Phase 5.12 bridge: eig_gsym_sub is Eigen.
-        helfem::Vector Ea_e; helfem::Matrix Ca_e;
-        scf::eig_gsym_sub(Ea_e, Ca_e, helfem::to_eigen(Fa), helfem::to_eigen(Sinvh), dsym);
-        Ea = helfem::to_arma(Ea_e);
-        Ca = helfem::to_arma(Ca_e);
-      }
-    else
-      { // Phase 5.11 bridge: eig_gsym is Eigen.
-        helfem::Vector Ea_e; helfem::Matrix Ca_e;
-        scf::eig_gsym(Ea_e, Ca_e, helfem::to_eigen(Fa), helfem::to_eigen(Sinvh));
-        Ea = helfem::to_arma(Ea_e);
-        Ca = helfem::to_arma(Ca_e);
-      }
-    // Enforce occupation according to specified symmetry
-    if(i<readocc) {
-      { // Phase 5.13 bridge: enforce_occupations is Eigen.
-        helfem::Matrix Ca_e = helfem::to_eigen(Ca);
-        helfem::Vector Ea_e = helfem::to_eigen(Ea);
-        scf::enforce_occupations(Ca_e, Ea_e, helfem::to_eigen(S), occnuma, occsym);
-        Ca = helfem::to_arma(Ca_e);
-        Ea = helfem::to_arma(Ea_e);
-      }
-    }
-
-    if(restr && nela==nelb) {
-      Eb=Ea;
-      Cb=Ca;
+    // Fock assembly. Restricted: one AO Fock matrix per block.
+    // Unrestricted: separate alpha/beta AO Fock matrices.
+    // Pre-assembled 1-electron core. Vel/Vmag are zero matrices when
+    // their coupling is off so this matches T + Vnuc + J exactly in the
+    // no-field case.
+    const arma::mat H1 = T + Vnuc + Vel + Vmag;
+    auto apply_mavg = [&](arma::mat & F) {
+      if (maverage)
+        F = helfem::to_arma(scf::fock_symmetry_average(helfem::to_eigen(F), mavg_idx));
+    };
+    if (restricted) {
+      arma::mat F_ao = H1 + J;
+      if (have_xc)  F_ao += XCa;
+      if (have_exx) F_ao += Ka;
+      apply_mavg(F_ao);
+      for (size_t k = 0; k < nsym; ++k)
+        orthonormalize_block(fock, k, F_ao, k);
     } else {
-      if(symm)
-        { // Phase 5.12 bridge: eig_gsym_sub is Eigen.
-          helfem::Vector Eb_e; helfem::Matrix Cb_e;
-          scf::eig_gsym_sub(Eb_e, Cb_e, helfem::to_eigen(Fb), helfem::to_eigen(Sinvh), dsym);
-          Eb = helfem::to_arma(Eb_e);
-          Cb = helfem::to_arma(Cb_e);
-        }
-      else
-        { // Phase 5.11 bridge: eig_gsym is Eigen.
-          helfem::Vector Eb_e; helfem::Matrix Cb_e;
-          scf::eig_gsym(Eb_e, Cb_e, helfem::to_eigen(Fb), helfem::to_eigen(Sinvh));
-          Eb = helfem::to_arma(Eb_e);
-          Cb = helfem::to_arma(Cb_e);
-        }
-    }
-    // Enforce occupation according to specified symmetry
-    if(i<readocc) {
-      { // Phase 5.13 bridge: enforce_occupations is Eigen.
-        helfem::Matrix Cb_e = helfem::to_eigen(Cb);
-        helfem::Vector Eb_e = helfem::to_eigen(Eb);
-        scf::enforce_occupations(Cb_e, Eb_e, helfem::to_eigen(S), occnumb, occsym);
-        Cb = helfem::to_arma(Cb_e);
-        Eb = helfem::to_arma(Eb_e);
+      arma::mat Fa_ao = H1 + J;
+      arma::mat Fb_ao = H1 + J;
+      if (have_xc)  { Fa_ao += XCa; Fb_ao += XCb; }
+      if (have_exx) { Fa_ao += Ka;  Fb_ao += Kb;  }
+      // Spin-Zeeman splitting Fa -= Bz/2 * S, Fb += Bz/2 * S. In
+      // restricted mode alpha == beta so no split is applied.
+      if (have_bfield) {
+        Fa_ao -= 0.5 * Bz * S;
+        Fb_ao += 0.5 * Bz * S;
+      }
+      apply_mavg(Fa_ao);
+      apply_mavg(Fb_ao);
+      for (size_t k = 0; k < nsym; ++k) {
+        orthonormalize_block(fock, k,        Fa_ao, k);
+        orthonormalize_block(fock, nsym + k, Fb_ao, k);
       }
     }
+    return std::make_pair(Etot, fock);
+  };
 
-    chkpt.write("Ca",Ca);
-    chkpt.write("Cb",Cb);
-    chkpt.write("Ea",Ea);
-    chkpt.write("Eb",Eb);
-
-    Caocc=Ca.cols(0,nela-1);
-    if(Ca.n_cols>(size_t) nela)
-      Cavirt=Ca.cols(nela,Ca.n_cols-1);
-    if(nelb>0)
-      Cbocc=Cb.cols(0,nelb-1);
-    if(Cb.n_cols>(size_t) nelb)
-      Cbvirt=Cb.cols(nelb,Cb.n_cols-1);
-    if(symm)
-      printf("Subspace diagonalization done in %.6f\n",timer.get());
-    else
-      printf("Full diagonalization done in %.6f\n",timer.get());
-
-    if(Ea.n_elem>(size_t)nela)
-      printf("Alpha HOMO-LUMO gap is % .3f eV\n",(Ea(nela)-Ea(nela-1))*HARTREEINEV);
-    if(nelb && Eb.n_elem>(size_t)nelb)
-      printf("Beta  HOMO-LUMO gap is % .3f eV\n",(Eb(nelb)-Eb(nelb-1))*HARTREEINEV);
-    fflush(stdout);
-
-    printf("\n");
-    printf("Alpha orbital symmetries\n");
-    classify_orbitals(Caocc,mvals,mposidx,mnegidx);
-    if(nelb>0) {
-      printf("\n");
-      printf("Beta orbital symmetries\n");
-      classify_orbitals(Cbocc,mvals,mposidx,mnegidx);
-    }
-    printf("\n");
-
-    if(convd)
-      break;
-  }
-
-  printf("%-21s energy: % .16f\n","Kinetic",Ekin);
-  printf("%-21s energy: % .16f\n","Nuclear attraction",Epot);
-  printf("%-21s energy: % .16f\n","Nuclear repulsion",Enucr);
-  printf("%-21s energy: % .16f\n","Coulomb",Ecoul);
-  printf("%-21s energy: % .16f\n","Exact exchange",Exx);
-  printf("%-21s energy: % .16f\n","Exchange-correlation",Exc);
-  printf("%-21s energy: % .16f\n","Electric field",Eefield);
-  printf("%-21s energy: % .16f\n","Magnetic field",Emfield);
-  printf("%-21s energy: % .16f\n","Nucleus-field",Enucfield);
-  printf("%-21s energy: % .16f\n","Total",Etot);
-  printf("%-21s energy: % .16f\n","Virial ratio",-Etot/Ekin);
-
-  printf("%-21s  force: %e\n", "Hellmann-Feynman", (2*Ekin+Epot+Enucr+Ecoul+Exx+Exc)/Rbond);
-
-  double eldip=-arma::trace(dip*P);
-  double elquad=-arma::trace(quad*P);
-
-  printf("\n");
-  printf("Electronic dipole     moment % .16e\n",eldip);
-  printf("Nuclear    dipole     moment % .16e\n",nucdip);
-  printf("Total      dipole     moment % .16e\n",eldip+nucdip);
-  printf("Electronic quadrupole moment % .16e\n",elquad);
-  printf("Nuclear    quadrupole moment % .16e\n",nucquad);
-  printf("Total      quadrupole moment % .16e\n",elquad+nucquad);
-
-  printf("\n");
-  printf("Nuclear electron densities\n");
-  arma::vec nucdena(basis.nuclear_density(Pa));
-  arma::vec nucdenb(basis.nuclear_density(Pb));
-  for(size_t i=0;i<nucdena.size();i++) {
-    int Z = (i==0) ? Z1 : Z2;
-    if(Z==0)
-      continue;
-    printf("%-2s: % .10e % .10e % .10e\n",element_symbols[Z].c_str(),nucdena(i),nucdenb(i),nucdena(i)+nucdenb(i));
-  }
-
-  // rms sizes
-  std::vector<arma::mat> orba, orbb;
-  for(int io=0;io<nela;io++)
-    orba.push_back(basis.radial_moments(Caocc.col(io)*arma::trans(Caocc.col(io))));
-  for(int io=0;io<nelb;io++)
-    orbb.push_back(basis.radial_moments(Cbocc.col(io)*arma::trans(Cbocc.col(io))));
-
-  for(int ic=0;ic<3;ic++) {
-    // Lh analysis only if lh atom exists
-    if(ic==0 && Z1==0)
-      continue;
-    // Middle analysis only if both atoms exist
-    if(ic==1 && (Z1==0 || Z2==0))
-      continue;
-    // Rh analysis only if rh atom exists
-    if(ic==2 && Z2==0)
-      continue;
-
-    if(ic==0)
-      printf("\nOccupied orbital analysis wrt left atom:\n");
-    else if(ic==1)
-      printf("\nOccupied orbital analysis wrt geometrical center:\n");
-    else
-      printf("\nOccupied orbital analysis wrt right atom:\n");
-
-    enum moment {mone,
-                 one,
-                 two,
-                 three};
-
-    if(ic==1) {
-      printf("Alpha orbitals\n");
-      printf("%2s %13s %12s\n","io","energy","sqrt(<r^2>)");
-      for(int io=0;io<nela;io++) {
-        printf("%2i % e %e\n",(int) io+1, Ea(io), sqrt(orba[io](two,ic)));
+  // Core-H guess per block per particle type. Fold in the external-field
+  // 1e matrices so the very first Fock reflects the perturbing potential.
+  const arma::mat H0 = T + Vnuc + Vel + Vmag;
+  OpenOrbitalOptimizer::FockMatrix<OOO_Real> CoreH(nsym * nparttype);
+  for (size_t t = 0; t < nparttype; ++t) {
+    for (size_t k = 0; k < nsym; ++k) {
+      if (!dsym[k].n_elem) {
+        CoreH[t * nsym + k] = helfem::Matrix::Zero(0, 0);
+        continue;
       }
-      printf("Beta orbitals\n");
-      for(int io=0;io<nelb;io++) {
-        printf("%2i % e %e\n",(int) io+1, Eb(io), sqrt(orbb[io](two,ic)));
-      }
-    } else {
-      printf("Alpha orbitals\n");
-      printf("%2s %13s %12s %12s %12s %12s\n","io","energy","1/<r^-1>","<r>","sqrt(<r^2>)","cbrt(<r^3>)");
-      for(int io=0;io<nela;io++) {
-        printf("%2i % e %e %e %e %e\n",(int) io+1, Ea(io), 1.0/orba[io](mone,ic), orba[io](one,ic), sqrt(orba[io](two,ic)), cbrt(orba[io](three,ic)));
-      }
-      printf("Beta orbitals\n");
-      for(int io=0;io<nelb;io++) {
-        printf("%2i % e %e %e %e %e\n",(int) io+1, Eb(io), 1.0/orbb[io](mone,ic), orbb[io](one,ic), sqrt(orbb[io](two,ic)), cbrt(orbb[io](three,ic)));
-      }
+      arma::mat H_sub = H0(dsym[k], dsym[k]);
+      if (have_bfield && nparttype == 2)
+        H_sub += (t == 0 ? -0.5 : 0.5) * Bz * arma::mat(S(dsym[k], dsym[k]));
+      const arma::mat H_orth = Sinvh_arma[k].t() * H_sub * Sinvh_arma[k];
+      CoreH[t * nsym + k] = helfem::to_eigen(H_orth);
     }
   }
 
-  /*
-  // Test orthonormality
-  arma::mat Smo(Ca.t()*S*Ca);
-  Smo-=arma::eye<arma::mat>(Smo.n_rows,Smo.n_cols);
-  printf("Alpha orthonormality deviation is %e\n",arma::norm(Smo,"fro"));
-  Smo=(Cb.t()*S*Cb);
-  Smo-=arma::eye<arma::mat>(Smo.n_rows,Smo.n_cols);
-  printf("Beta orthonormality deviation is %e\n",arma::norm(Smo,"fro"));
-  */
+  OpenOrbitalOptimizer::SCFSolver<OOO_Real, OOO_Real> scfsolver(
+      number_of_blocks_per_particle_type, maximum_occupation,
+      number_of_particles, fock_builder, block_descriptions);
+
+  // --readocc: parse occs.dat and hand OOO a fixed per-block particle
+  // count. Bespoke diatomic reads (nocca, noccb, m) rows for both
+  // hetero- and homonuclear cases (symm=0/1), and adds a fourth
+  // `parity in {+1,-1}` column when symm=2 (homonuclear g/u split).
+  //
+  // Same bool-vs-int semantic difference as atomic_ooo: OOO's frozen
+  // per-block particles cannot be released mid-SCF, so we freeze for
+  // the entire run instead of the first N Fock builds only.
+  if (readocc) {
+    if (symm == 0)
+      throw std::logic_error("--readocc requires --symmetry>=1 (need per-block index).");
+    arma::imat occs;
+    occs.load("occs.dat", arma::raw_ascii);
+    if (Z1 != Z2 && occs.n_cols != 3)
+      throw std::logic_error("occs.dat: heteronuclear molecule requires 3 columns (nocca, noccb, m).");
+    if (Z1 == Z2 && occs.n_cols != 3 && occs.n_cols != 4)
+      throw std::logic_error("occs.dat: homonuclear molecule requires 3 or 4 columns.");
+    if (occs.n_cols == 4 && symm != 2)
+      throw std::logic_error("occs.dat: 4-column form (with parity) requires --symmetry=2.");
+
+    Eigen::Matrix<OOO_Real, Eigen::Dynamic, 1> fixed_particles =
+        Eigen::Matrix<OOO_Real, Eigen::Dynamic, 1>::Zero(nsym * nparttype);
+    int total_a = 0, total_b = 0;
+    for (arma::uword i = 0; i < occs.n_rows; ++i) {
+      const int nocca_i = static_cast<int>(occs(i, 0));
+      const int noccb_i = static_cast<int>(occs(i, 1));
+      arma::uvec row_idx;
+      if (occs.n_cols == 3) {
+        row_idx = basis.m_indices(occs(i, 2));
+      } else {
+        if (occs(i, 3) != 1 && occs(i, 3) != -1)
+          throw std::logic_error("occs.dat: parity column must be +1 or -1.");
+        row_idx = basis.m_indices(occs(i, 2), (occs(i, 3) == -1));
+      }
+      if (!row_idx.n_elem)
+        throw std::logic_error("occs.dat: row references a symmetry block with no basis functions.");
+      int matched_block = -1;
+      for (size_t k = 0; k < nsym; ++k) {
+        if (dsym[k].n_elem == row_idx.n_elem &&
+            arma::all(dsym[k] == row_idx)) { matched_block = static_cast<int>(k); break; }
+      }
+      if (matched_block < 0)
+        throw std::logic_error("occs.dat: row does not match any current symmetry block.");
+      total_a += nocca_i;
+      total_b += noccb_i;
+      if (restricted) {
+        if (nocca_i != noccb_i)
+          throw std::logic_error("occs.dat: nocca != noccb on a row is incompatible with restricted mode.");
+        fixed_particles(matched_block) = static_cast<OOO_Real>(nocca_i + noccb_i);
+      } else {
+        fixed_particles(matched_block)         = static_cast<OOO_Real>(nocca_i);
+        fixed_particles(nsym + matched_block)  = static_cast<OOO_Real>(noccb_i);
+      }
+    }
+    if (total_a != nela)
+      throw std::logic_error("occs.dat: sum of nocca does not match --nela.");
+    if (total_b != nelb)
+      throw std::logic_error("occs.dat: sum of noccb does not match --nelb.");
+    scfsolver.fixed_number_of_particles_per_block(fixed_particles);
+  }
+
+  scfsolver.initialize_with_fock(CoreH);
+  scfsolver.run();
 
   return 0;
-
 }
